Fix boundaries code generation

+++ python/pystencils_walberla/boundary.py
@@ -97,7 +97,7 @@ def generate_boundary(generation_context,
         'inner_or_boundary': boundary_object.inner_or_boundary,
         'single_link': boundary_object.single_link,
         'additional_data_handler': additional_data_handler,
-        'dtype': "double" if is_float else "float",
+        'dtype': "float" if is_float else "double",
         'layout': layout
     }
